requirements.txt
pandas
Faker
mysql-connector-python
sqlalchemy
